Move BlobFileAddition and BlobFileGarbage to their own separate source files
