+++ CMakeLists.txt
@@ -501,7 +501,8 @@ set(SOURCES
         cache/lru_cache.cc
         cache/sharded_cache.cc
         db/arena_wrapped_db_iter.cc
-        db/blob_file_state.cc
+        db/blob_file_addition.cc
+        db/blob_file_garbage.cc
         db/builder.cc
         db/c.cc
         db/column_family.cc
@@ -925,7 +926,8 @@ if(WITH_TESTS)
   set(TESTS
         cache/cache_test.cc
         cache/lru_cache_test.cc
-        db/blob_file_state_test.cc
+        db/blob_file_addition_test.cc
+        db/blob_file_garbage_test.cc
         db/column_family_test.cc
         db/compact_files_test.cc
         db/compaction/compaction_job_stats_test.cc
