+++ TARGETS
@@ -116,7 +116,8 @@ cpp_library(
         "cache/lru_cache.cc",
         "cache/sharded_cache.cc",
         "db/arena_wrapped_db_iter.cc",
-        "db/blob_file_state.cc",
+        "db/blob_file_addition.cc",
+        "db/blob_file_garbage.cc",
         "db/builder.cc",
         "db/c.cc",
         "db/column_family.cc",
@@ -480,8 +481,15 @@ ROCKS_TESTS = [
         [],
     ],
     [
-        "blob_file_state_test",
-        "db/blob_file_state_test.cc",
+        "blob_file_addition_test",
+        "db/blob_file_addition_test.cc",
+        "serial",
+        [],
+        [],
+    ],
+    [
+        "blob_file_garbage_test",
+        "db/blob_file_garbage_test.cc",
         "serial",
         [],
         [],
